requirements/develop.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
Django==4.1.7
django-cors-headers==3.14.0
django-redis==5.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg==3.1.8
psycopg-binary==3.1.8
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.2
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
